refactor: streamline ATS components and enhance candidate data handling
- Removed outdated ATS page components and consolidated candidate data management into a more efficient structure.
- Introduced a new `ATSCandidatesTable` component for improved rendering and sorting of candidate data.
- Updated API interactions to ensure consistent data fetching and error handling.
- Enhanced the dashboard layout with responsive design elements and integrated ATS access control checks.
- Improved user experience by implementing loading states and error messages for better feedback during data operations.
